Fix kembalian calculation, support decimal item quantity, and update thermal receipt layout

+++ app/Models/SaleItem.php
@@ -22,7 +22,7 @@ class SaleItem extends Model
     {
         return [
             'price' => 'decimal:2',
-            'quantity' => 'integer',
+            'quantity' => 'float',
             'subtotal' => 'decimal:2',
         ];
     }
